Put MatrixField in matrix.py
