+++ python/taichi/aot/module.py
@@ -1,7 +1,8 @@
 from contextlib import contextmanager
 
 from taichi.lang import impl, kernel_arguments, kernel_impl
-from taichi.lang.field import MatrixField, ScalarField
+from taichi.lang.field import ScalarField
+from taichi.lang.matrix import MatrixField
 
 
 class KernelTemplate:
