+++ python/taichi/lang/impl.py
@@ -6,7 +6,8 @@ import numpy as np
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang.exception import InvalidOperationError, TaichiSyntaxError
 from taichi.lang.expr import Expr, make_expr_group
-from taichi.lang.field import Field, MatrixField, ScalarField
+from taichi.lang.field import Field, ScalarField
+from taichi.lang.matrix import MatrixField
 from taichi.lang.snode import SNode
 from taichi.lang.tape import TapeImpl
 from taichi.lang.util import (cook_dtype, is_taichi_class, python_scope,
